model2/requirements.txt
flask
langchain
PyPDF2
langchain_google_genai
langchain_community
faiss-cpu
streamlit
requests
python-dotenv
